moved fixtures
