+++ tests/conftest.py
@@ -564,6 +564,45 @@ def total_level_df_w_nans():
 
 
 @pytest.fixture
+def product_level_constant_hierarchical_df():
+    df = pd.DataFrame(
+        {
+            "timestamp": ["2000-01-01", "2000-01-02", "2000-01-03", "2000-01-04"] * 4,
+            "segment": ["a"] * 4 + ["b"] * 4 + ["c"] * 4 + ["d"] * 4,
+            "target": [1, 1, 1, 1] + [2, 2, 2, 2] + [3, 3, 3, 3] + [4, 4, 4, 4],
+        }
+    )
+    df = TSDataset.to_dataset(df)
+    return df
+
+
+@pytest.fixture
+def market_level_constant_hierarchical_df():
+    df = pd.DataFrame(
+        {
+            "timestamp": ["2000-01-01", "2000-01-02", "2000-01-03", "2000-01-04"] * 2,
+            "segment": ["X"] * 4 + ["Y"] * 4,
+            "target": [3, 3, 3, 3] + [7, 7, 7, 7],
+        }
+    )
+    df = TSDataset.to_dataset(df)
+    return df
+
+
+@pytest.fixture
+def market_level_constant_hierarchical_df_exog():
+    df = pd.DataFrame(
+        {
+            "timestamp": ["2000-01-01", "2000-01-02", "2000-01-03", "2000-01-04", "2000-01-05", "2000-01-06"] * 2,
+            "segment": ["X"] * 6 + ["Y"] * 6,
+            "regressor": [1, 1, 1, 1, 1, 1] * 2,
+        }
+    )
+    df = TSDataset.to_dataset(df)
+    return df
+
+
+@pytest.fixture
 def total_level_simple_hierarchical_ts(total_level_df, hierarchical_structure):
     ts = TSDataset(df=total_level_df, freq="D", hierarchical_structure=hierarchical_structure)
     return ts
@@ -584,4 +623,80 @@ def product_level_simple_hierarchical_ts(product_level_df, hierarchical_structur
 @pytest.fixture
 def simple_no_hierarchy_ts(market_level_df):
     ts = TSDataset(df=market_level_df, freq="D")
+    return ts
+
+
+@pytest.fixture
+def market_level_constant_hierarchical_ts(market_level_constant_hierarchical_df, hierarchical_structure):
+    ts = TSDataset(df=market_level_constant_hierarchical_df, freq="D", hierarchical_structure=hierarchical_structure)
+    return ts
+
+
+@pytest.fixture
+def market_level_constant_hierarchical_ts_w_exog(
+    market_level_constant_hierarchical_df, market_level_constant_hierarchical_df_exog, hierarchical_structure
+):
+    ts = TSDataset(
+        df=market_level_constant_hierarchical_df,
+        df_exog=market_level_constant_hierarchical_df_exog,
+        freq="D",
+        hierarchical_structure=hierarchical_structure,
+        known_future="all",
+    )
+    return ts
+
+
+@pytest.fixture
+def product_level_constant_hierarchical_ts(product_level_constant_hierarchical_df, hierarchical_structure):
+    ts = TSDataset(
+        df=product_level_constant_hierarchical_df,
+        freq="D",
+        hierarchical_structure=hierarchical_structure,
+    )
+    return ts
+
+
+@pytest.fixture
+def product_level_constant_hierarchical_ts_w_exog(
+    product_level_constant_hierarchical_df, market_level_constant_hierarchical_df_exog, hierarchical_structure
+):
+    ts = TSDataset(
+        df=product_level_constant_hierarchical_df,
+        df_exog=market_level_constant_hierarchical_df_exog,
+        freq="D",
+        hierarchical_structure=hierarchical_structure,
+        known_future="all",
+    )
+    return ts
+
+
+@pytest.fixture
+def product_level_constant_forecast_w_quantiles(hierarchical_structure):
+    df = pd.DataFrame(
+        {
+            "timestamp": ["2000-01-05", "2000-01-06"] * 4,
+            "segment": ["a"] * 2 + ["b"] * 2 + ["c"] * 2 + ["d"] * 2,
+            "target": [1, 1] + [2, 2] + [3, 3] + [4, 4],
+            "target_0.25": [1, 1] + [2, 2] + [3, 3] + [4, 4],
+            "target_0.75": [1, 1] + [2, 2] + [3, 3] + [4, 4],
+        }
+    )
+    df = TSDataset.to_dataset(df=df)
+    ts = TSDataset(df=df, freq="D", hierarchical_structure=hierarchical_structure)
+    return ts
+
+
+@pytest.fixture
+def total_level_constant_forecast_w_quantiles(hierarchical_structure):
+    df = pd.DataFrame(
+        {
+            "timestamp": ["2000-01-05", "2000-01-06"],
+            "segment": ["total"] * 2,
+            "target": [10, 10],
+            "target_0.25": [10, 10],
+            "target_0.75": [10, 10],
+        }
+    )
+    df = TSDataset.to_dataset(df=df)
+    ts = TSDataset(df=df, freq="D", hierarchical_structure=hierarchical_structure)
     return ts
